Adding another patch method
