+++ users/michaelt/bitlon_tweet/run.py
@@ -4,7 +4,8 @@ from datetime import datetime, timedelta
 import twitter
 import pytz
 
-from users.michaelt.bitlon_tweet import scheduler, report, get_data, api_patch
+from users.michaelt.bitlon_tweet import scheduler, report, get_data
+from api_patches.coinapi_patch import CoinAPIv1Patched
 
 def main():
     twitter_api = twitter.Api(consumer_key=os.environ['TWITTER_API_KEY'],
@@ -13,7 +14,7 @@ def main():
                           access_token_secret=os.environ['TWITTER_ACCESS_TOKEN_SECRET']
     )
 
-    coinapi = api_patch.CoinAPIv1Patched(os.environ['COINAPI_API_KEY'])
+    coinapi = CoinAPIv1Patched(os.environ['COINAPI_API_KEY'])
     utc = pytz.UTC
 
     tweets = get_data.get_user_latest_tweets(
